+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - PR [337](https://github.com/plugwise/python-plugwise-usb/pull/337): Improve node removal, remove and reset the node as executed by Source, and remove the cache-file.
 - PR [342](https://github.com/plugwise/python-plugwise-usb/pull/342): Improve node_type chaching.
+- PR [343](https://github.com/plugwise/python-plugwise-usb/pull/343): Improve writing of cache-files.
 
 ## 0.46.0 - 2025-09-12